+++ eblox/lib/eblox.dart
@@ -4,3 +4,5 @@ export 'src/blox_base.dart';
 export 'src/blox_provider.dart';
 export 'src/blox_view.dart';
 export 'src/blox_data.dart';
+export 'src/blox_action.dart';
+export 'src/blox_state.dart';