+++ eblox/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.2
+- Add `BloxAsyncTask` type as return value.
+
+
 ## 0.0.1
 
 * TODO: Describe initial release.