+++ eblox_annotation/CHANGELOG.md
@@ -1,3 +1,7 @@
+
+## 0.0.2
+- Added `@bindAsync` annotation
+
 ## 0.0.1
 
 - Initial version.